完善宿主项目的 cake 构建脚本，增加标记参数与相关条件参数。 🍰

+++ terminal/build.cake
@@ -1,14 +1,26 @@
-var scheme = Argument("scheme", "default");
+[Flags]
+enum Flags
+{
+	None = 0,
+	NoClean = 1,
+	NoRestore = 2,
+}
+
+var flags = Argument("flags", Flags.None);
 var target = Argument("target", "default");
 var edition = Argument("edition", "Debug");
-var framework = Argument("framework", "net9.0");
+var platform = Argument("platform", "Windows");
+var framework = Argument("framework", "net10.0");
 var architecture = Argument("architecture", "x64");
-var solutionFile  = "Zongsoft.Hosting.Terminal.slnx";
+var solutionFile = "Zongsoft.Hosting.Terminal.slnx";
 
 Task("clean")
 	.Description("清理解决方案")
 	.Does(() =>
 {
+	if((flags & Flags.NoClean) == Flags.NoClean)
+		return;
+
 	DeleteFiles("*.nupkg");
 	CleanDirectories("**/bin");
 	CleanDirectories("**/obj");
@@ -18,7 +30,22 @@ Task("restore")
 	.Description("还原项目依赖")
 	.Does(() =>
 {
-	DotNetRestore(solutionFile);
+	if((flags & Flags.NoRestore) == Flags.NoRestore)
+		return;
+
+	var settings = new DotNetRestoreSettings
+	{
+		MSBuildSettings = new DotNetMSBuildSettings()
+	};
+
+	if(string.Equals(platform, "win", StringComparison.OrdinalIgnoreCase))
+		platform = "Windows";
+
+	if(!string.IsNullOrEmpty(platform))
+		settings.MSBuildSettings.WithProperty("DefineConstants", platform.ToUpperInvariant());
+
+	settings.MSBuildSettings.WithProperty("Platform", "Any CPU");
+	DotNetRestore(solutionFile, settings);
 });
 
 Task("build")
@@ -29,17 +56,19 @@ Task("build")
 {
 	var settings = new DotNetBuildSettings
 	{
-		NoRestore = true
+		NoRestore = true,
+		Configuration = edition,
+		MSBuildSettings = new DotNetMSBuildSettings()
 	};
 
-	DotNetBuild(solutionFile, settings);
-});
+	if(string.Equals(platform, "win", StringComparison.OrdinalIgnoreCase))
+		platform = "Windows";
 
-Task("deploy")
-	.Description("部署插件")
-	.Does(() =>
-{
-	DotNetTool(solutionFile, "deploy", $" -host:terminal -site:daemon -scheme:{scheme} -edition:{edition} -framework:{framework} -architecture:{architecture} -verbosity:quiet");
+	if(!string.IsNullOrEmpty(platform))
+		settings.MSBuildSettings.WithProperty("DefineConstants", platform.ToUpperInvariant());
+
+	settings.MSBuildSettings.WithProperty("Platform", "Any CPU");
+	DotNetBuild(solutionFile, settings);
 });
 
 Task("default")
